Started development of component cxx chipset dma.
git-svn-id: https://unisim.org/svn/devel@5168 ea533459-3c0a-0410-a024-a90cf4b01ff5

+++ unisim/component/cxx/chipset/mpc107/dma/buffer_64.cc
@@ -44,7 +44,7 @@ namespace mpc107 {
 namespace dma {
 
 template
-class Buffer<uint32_t>;
+class Buffer<uint64_t>;
 
 } // end of namespace dma
 } // end of namespace mpc107
